# src/main/java/com/example/bankcards/dto/UserEnabledUpdateRequest.java
package com.example.bankcards.dto;

import jakarta.validation.constraints.NotNull;

public record UserEnabledUpdateRequest(
        @NotNull(message = "enabled must not be null")
        Boolean enabled
) {}